moved data factory into Authenticate, split it up to reduce database stress.
